refactor: extract shared NostrConnectForm from login page
Move Nostr login UI (Private key / Bunker tabs + Extension button)
into a reusable NostrConnectForm component. Admin login page and
landing claim dialog both consume it with customizable labels and
onSuccess callbacks.
